Réduction en mémoire (pawn = byte)

+++ src/model/QuartoModel.java
@@ -36,7 +36,7 @@ public class QuartoModel {
         playerType[0] = firstPlayerType;
         playerType[1] = secondPlayerType;
         pawnAvailable = new QuartoPawn[16];
-        for (int count = 0; count < 16; count++) {
+        for (byte count = 0; count < 16; count++) {
             pawnAvailable[count] = new QuartoPawn(count);
         }
     }
